Converted to maven project. Added 2 plugins jar/shade. ran mvn clean install to create the jar file in target.
