--- src/main/java/LinkedList/Node.java
@@ -1,12 +1,0 @@
-package LinkedList;
-
-public class Node<T> {
-
-    T data;
-    Node<T> next;
-
-    public Node(T data) {
-        this.data = data;
-        this.next = null;
-    }
-}
