--- src/main/java/airport/CarryOnBag.java
@@ -1,11 +1,0 @@
-package airport;
-
-public class CarryOnBag extends Luggage {
-
-    private final String LUGGAGE_TYPE = "CarryOn";
-
-    public CarryOnBag(int bagCount, int weight) {
-        super(bagCount, weight);
-    }
-
-}
