+++ src/main/java/com/solvd/airport/Airline.java
@@ -1,0 +1,24 @@
+package com.solvd.airport;
+
+public class Airline {
+
+    private String companyName;
+
+    public Airline(String airlineCompany) {
+        this.companyName = airlineCompany;
+    }
+
+    public String getName() {
+        return this.companyName;
+    }
+
+
+    public void setCompanyName(String company) {
+        this.companyName = company;
+    }
+
+    @Override
+    public String toString() {
+        return "The airport.Airline is " + this.companyName;
+    }
+}
