+++ src/main/java/com/solvd/airport/CarryOnBag.java
@@ -1,0 +1,11 @@
+package com.solvd.airport;
+
+public class CarryOnBag extends Luggage {
+
+    private final String LUGGAGE_TYPE = "CarryOn";
+
+    public CarryOnBag(int bagCount, int weight) {
+        super(bagCount, weight);
+    }
+
+}
