+++ src/main/java/com/solvd/exceptions/BagOverweightException.java
@@ -1,0 +1,7 @@
+package com.solvd.exceptions;
+
+public class BagOverweightException extends Exception {
+    public BagOverweightException(String message) {
+        super(message);
+    }
+}
