+++ src/main/java/com/solvd/exceptions/FlightNotFoundException.java
@@ -1,0 +1,7 @@
+package com.solvd.exceptions;
+
+public class FlightNotFoundException extends Exception {
+    public FlightNotFoundException(String message) {
+        super(message);
+    }
+}
