+++ src/main/java/com/solvd/exceptions/InvalidFoodItemException.java
@@ -1,0 +1,7 @@
+package com.solvd.exceptions;
+
+public class InvalidFoodItemException extends Exception {
+    public InvalidFoodItemException(String message) {
+        super(message);
+    }
+}
