+++ src/main/java/com/solvd/exceptions/InvalidGateException.java
@@ -1,0 +1,7 @@
+package com.solvd.exceptions;
+
+public class InvalidGateException extends Exception {
+    public InvalidGateException(String message) {
+        super(message);
+    }
+}
