+++ src/main/java/com/solvd/exceptions/PassengerNotFoundException.java
@@ -1,0 +1,7 @@
+package com.solvd.exceptions;
+
+public class PassengerNotFoundException extends Exception {
+    public PassengerNotFoundException(String message) {
+        super(message);
+    }
+}
