+++ src/main/java/com/solvd/exceptions/RestroomUnavailableException.java
@@ -1,0 +1,7 @@
+package com.solvd.exceptions;
+
+public class RestroomUnavailableException extends Exception {
+    public RestroomUnavailableException(String message) {
+        super(message);
+    }
+}
