+++ src/main/java/com/solvd/interfaces/IAnnouncement.java
@@ -1,0 +1,11 @@
+package com.solvd.interfaces;
+
+public interface IAnnouncement {
+
+    void gateAnnouncement();
+
+    void lotAnnouncement();
+
+    void employeeAnnouncement();
+
+}
