+++ src/main/java/com/solvd/interfaces/IEat.java
@@ -1,0 +1,8 @@
+package com.solvd.interfaces;
+
+import com.solvd.airport.FoodCourt;
+
+public interface IEat {
+
+    void eat(FoodCourt foodCourt);
+}
