+++ src/main/java/com/solvd/interfaces/IFacility.java
@@ -1,0 +1,7 @@
+package com.solvd.interfaces;
+
+public interface IFacility {
+
+    String getName();
+
+}
