+++ src/main/java/com/solvd/interfaces/ILuggageWeightCheck.java
@@ -1,0 +1,8 @@
+package com.solvd.interfaces;
+
+import com.solvd.exceptions.BagOverweightException;
+
+public interface ILuggageWeightCheck {
+
+    void weightCheck() throws BagOverweightException;
+}
