+++ src/main/java/com/solvd/interfaces/IPlane.java
@@ -1,0 +1,12 @@
+package com.solvd.interfaces;
+
+public interface IPlane {
+
+    void arrive();
+
+    void printBoardingTime();
+
+    void boardingStatus();
+
+    void depart();
+}
