+++ src/main/java/com/solvd/interfaces/IRestroomAvailability.java
@@ -1,0 +1,7 @@
+package com.solvd.interfaces;
+
+public interface IRestroomAvailability {
+
+    void checkAvailability();
+
+}
