+++ src/main/java/com/solvd/linkedlist/Node.java
@@ -1,0 +1,12 @@
+package com.solvd.linkedlist;
+
+public class Node<T> {
+
+    T data;
+    Node<T> next;
+
+    public Node(T data) {
+        this.data = data;
+        this.next = null;
+    }
+}
