--- src/main/java/exceptions/BagOverweightException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class BagOverweightException extends Exception {
-    public BagOverweightException(String message) {
-        super(message);
-    }
-}
