--- src/main/java/exceptions/FlightNotFoundException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class FlightNotFoundException extends Exception {
-    public FlightNotFoundException(String message) {
-        super(message);
-    }
-}
