--- src/main/java/exceptions/InvalidFoodItemException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidFoodItemException extends Exception {
-    public InvalidFoodItemException(String message) {
-        super(message);
-    }
-}
