--- src/main/java/exceptions/InvalidGateException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidGateException extends Exception {
-    public InvalidGateException(String message) {
-        super(message);
-    }
-}
