--- src/main/java/exceptions/OverWeightBagException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class OverWeightBagException extends Exception {
-    public OverWeightBagException(String message) {
-        super(message);
-    }
-}
