--- src/main/java/exceptions/PassengerNotFoundException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class PassengerNotFoundException extends Exception {
-    public PassengerNotFoundException(String message) {
-        super(message);
-    }
-}
