--- src/main/java/exceptions/RestroomUnavailableException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class RestroomUnavailableException extends Exception {
-    public RestroomUnavailableException(String message) {
-        super(message);
-    }
-}
