--- src/main/java/interfaces/IAnnouncement.java
@@ -1,11 +1,0 @@
-package interfaces;
-
-public interface IAnnouncement {
-
-    void gateAnnouncement();
-
-    void lotAnnouncement();
-
-    void employeeAnnouncement();
-
-}
