--- src/main/java/interfaces/IEat.java
@@ -1,8 +1,0 @@
-package interfaces;
-
-import airport.FoodCourt;
-
-public interface IEat {
-
-    void eat(FoodCourt foodCourt);
-}
