--- src/main/java/interfaces/IFacility.java
@@ -1,7 +1,0 @@
-package interfaces;
-
-public interface IFacility {
-
-    String getName();
-
-}
