--- src/main/java/interfaces/ILuggageWeightCheck.java
@@ -1,8 +1,0 @@
-package interfaces;
-
-import exceptions.BagOverweightException;
-
-public interface ILuggageWeightCheck {
-
-    void weightCheck() throws BagOverweightException;
-}
