--- src/main/java/interfaces/IPlane.java
@@ -1,12 +1,0 @@
-package interfaces;
-
-public interface IPlane {
-
-    void arrive();
-
-    void boardingTime();
-
-    void boardingStatus();
-
-    void depart();
-}
