--- src/main/java/interfaces/IRestroomAvailability.java
@@ -1,7 +1,0 @@
-package interfaces;
-
-public interface IRestroomAvailability {
-
-    void checkAvailability();
-
-}
